move admin privileges to admin page
