+++ lib/quadquizaminos_web/live/admin_contests_live.html.leex
@@ -1,12 +1,11 @@
     <div class="container">
-     <%= if @live_action == :show do %>
-     <%= live_component @socket, QuadquizaminosWeb.ContestFinalResultComponent, contest_records: @contest_records, id: "final_result" %>
-     <% else %>
- <%= if @countdown_interval <= 0  do %>
         <h1>Contests</h1>
+         <input type="text" phx-keydown="save"  phx-key="Enter">
              <table>
              <tr>
                  <th>Name</th>
+                 <th></th>
+                 <th></th>
                  <th></th>
                  <th>Start time</th>
                  <th>End time</th>
@@ -16,6 +15,12 @@
                   <tr>
                   <td><%= contest.name%></td>
                  <td>
+                 <%= start_or_pause_button(assigns,contest) %>
+                 </td>
+                 <td>
+                     <button class="<%= maybe_disable_button(contest) %> <%= if contest_running?(contest), do: "red" %> icon-button" phx-click="stop" phx-value-contest='<%= contest.name %>' <%= maybe_disable_button(contest) %>><i class="fas fa-stop-circle fa-2x"></i></button>
+                 </td>
+                 <td>
                 <%= timer_or_final_result(assigns, contest) %>
                 </td>
                 <td><%= truncate_date(contest.start_time) %></td>
@@ -24,9 +29,5 @@
             </tr>
           <% end %>
         </table>
-    <% else %>
-        <%= countdown_timer(assigns) %>
-     <% end %>
         
-<% end %>
 </div>
